requirements.txt
django==4.1.4
environs[django]==9.5.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
pillow==9.3.0
django-filter==22.1